requirement.txt
asgiref==3.8.1
attrs==24.2.0
distlib==0.3.8
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.6.1
drf-yasg==1.21.7
filelock==3.15.4
importlib_resources==6.4.5
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.1
platformdirs==4.2.2
psycopg2==2.9.9
PyJWT==2.9.0
pymongo==4.8.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
swagger-spec-validator==3.0.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
virtualenv==20.26.3
virtualenvwrapper-win==1.2.7
whitenoise==6.7.0
